Always install the object in "call_graph" schema.
Supporting relocation seemed to be more trouble than worth at the
moment, so just force the schema to "call_graph".

Also get rid of the pgcrypto dependency by using the built-in md5()
for the hashing instead.  This is mostly because supporting pgcrypto
in an arbitrary schema seemed hard to do. :-(