Fix enabling call_graph mid-flight
Previously we didn't keep any track of the call graph when the
module wasn't enabled.  This, however, meant that enabling call_graph
on a running system might produce faulty data, or even crashes because
we might start tracking only a branch of a call graph thinking we were
tracking all of it.

Instead, when the module is disabled, only keep track of how many times
we've recursed into the top level function to know when we're done
with a single call graph so we can start tracking the next one.  This
works so long as the module is loaded in shared_preload_libraries.