Change references of fake_pgproc to pg_proc
I seem to have forgot to change this before I committed
48f49a99aa32b7de120338a2d12cf3ab80ada63d which introduced
create_graphs.pl.

The reason the script had fake_pgproc in the first place is that
it was a hack which allowed me to create graphs inside a database
other than the one that originally tracked the call graphs.  Maybe
in the future it would make sense to allow the user to specify a
table (similar to my fake_pgproc) to dig the function names from,
but for now, this should do..