Add more information about the graphs on the HTML page
Using just a single query and CTEs wasn't going to cut it, so I had
to split the CTEs into temporary tables.  I did keep the nodes and
edges in the same query (at least for the time being), because
I didn't see any benefit from splitting it.  To the contrary, it
would have made the dot file loop's logic a bit more complex.