Add GeneratePerFunctionGraphs configuration parameter
This also means that calling generate_per_function_graphs.pl
(previously create_per_function_graphs.pl) is not callable from the
command line anymore.

Because the data type of the parameter is effectively boolean, we
need some special code to handle it (similar to SubGraphs, which is
an array).  For now, special casing these parameters will work..

Also contains a number of changes to keep the overall coding style
a bit more consistent.