Allow table usage tracking
This new feature allows the user to track which tables are used by
the different call graphs.  Call graphs are still identified by the
hash of their edges, which means that you can later add table data
to the call graphs tracked with track_table_usage=off.

Unfortunately because there is no way to reset the table usage
statistics in the middle of a transaction, there is no way to track
the table usage information reliably in multi-statement transactions.
To avoid giving false information to the user, we have no other option
than to silently discard any table usage data if the top level
function call was not the first statement in the transaction.  We
could remove this restriction by taking snapshots of
pg_stat_xact_user_tables between top level function calls, but that
didn't seem worth the trouble for our use case.  Also note that
BEGIN counts as a statement, so a sequence like  BEGIN; SELECT f();
COMMIT;  would not track table usage data either.

Currently there is no support for creating visual graphs out of this
data, but that will be implemented in the near future.

Also bump the version number to 1.1 and provide upgrade scripts from
1.0 to 1.1.