Fix performance problems in ProcessCallGraphBuffers()
This commit includes two performance changes:
  1) remove an unnecessary WHERE .. ANY (_.CallGraphIDs)
       This for some reason made the CTE scan crawl on its knees.
       But the entire check was unnecessary anyway (since we check
       the IDs in the CTE), so simply remove the check.
  2) process at most _MaxBufferCount buffers on a single call
       This sets an upper bound on the size of the _.CallGraphIDs
       array, which seems to improve things immensely.  Also add an
       index on (CallGraphBufferID, TopLevelFunction) to remove the
       need to re-sort the output of the "Buffers" subquery in
       ProcessCallGraphBuffers().

Also a small change to the return value of the function: instead of
returning the number of call graphs that changed, return an actual
list of call graph ids for the affected graphs.  This allows us to
(in the future) only regenerate visualizations of graphs that
changed, instead of always regenerating all of them.

Schema version bump to 1.2, not yet tagged because of virtually no
testing. :-(