Fix wrong column name in call_graph.c
The column is actually called idx_tup_fetch, not idx_tup_read.  I'm
sure I tested this, so I don't know how this happened.. :-(

It's a bit shame that the columns in call_graph.* tables are now
all called idx_tup_read, but I think we'll have to live with that.
I don't want to do another schema version bump.