Move subgraphs to a separate HTML file
This looks a lot better, but also makes the code even more complex.
:-(

Also includes a small refactoring to create_graphs.pl: I didn't like
the way the IsSubGraph and EntryFunctionOid information was parsed
from the GraphID, so I added extra columns for both of them.

I originally wanted to change the name of the "main graphs" from
"graph1" to EntryFunctionName, but that seemed a bit harder to do.
Maybe some other day..