Rewrite table usage tracking
After some testing it occurred that the feature introduced in
6071135ca3e030da5a43b12880b3d9a350f567ad was not enough even for
our purposes since EXECUTE prepared_statement(); counts as two
statements, and thus will not track table information. :-(

The way to fix this was to take a snapshot of pg_stat_act_user_tables
before entering the top level function, and then comparing that to
the data in the table before leaving the top level function.  This,
however, comes with a significant performance penalty, and I don't
see any obvious way to make things significantly better without
modifying PG itself.  One low-hanging fruit would be to avoid using
the pg_stat_xact_*_tables views and do the hash lookups ourselves,
but I'm not sure how much that would help.  Maybe in the future we
could look into changing things in PG, but for now, this is all we
got..