Hide pg_stat_xact_user_tables information from TLF graphs
Now that table usage tracking works for all callers, the graphs
quickly grew bigger than necessary.  Now the only information we
show is the name of the table and whether it was written to, or
only read from.  UPDATE and DELETE both require a seq/idx scan
before they can modify the data so we can't realiably tell the user
that "this table is only written to"; it would only work for
INSERTs.